Move event checkin from checkin to event service
